requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
discord.py==1.7.2
idna==2.10
multidict==5.1.0
requests==2.25.1
tinydb==4.4.0
typing-extensions==3.10.0.0
urllib3==1.26.4
yarl==1.6.3
